Possibly more type fixes.

+++ main/src/addins/VersionControl/MonoDevelop.VersionControl.Subversion.Unix/MonoDevelop.VersionControl.Subversion.Unix/LibSvnClient.cs
@@ -454,7 +454,7 @@ namespace MonoDevelop.VersionControl.Subversion.Unix {
 			public IntPtr repos_lock;
 			public IntPtr url;
 			public svn_revnum_t ood_last_cmt_rev;
-			public long ood_last_cmt_date;
+			public IntPtr ood_last_cmt_date;
 			public svn_node_kind_t ood_kind;
 			public string ood_last_cmt_author;
 			public IntPtr tree_conflict;
@@ -559,7 +559,7 @@ namespace MonoDevelop.VersionControl.Subversion.Unix {
 		[StructLayout (LayoutKind.Sequential)]
 		public struct svn_auth_cred_ssl_server_trust_t {
 			[MarshalAs (UnmanagedType.Bool)] public bool may_save;
-			public uint accepted_failures;
+			public UInt32 accepted_failures;
 		}
 
 		[StructLayout (LayoutKind.Sequential)]
@@ -655,21 +655,20 @@ namespace MonoDevelop.VersionControl.Subversion.Unix {
 		
 		public delegate IntPtr svn_auth_simple_prompt_func_t (ref IntPtr cred, IntPtr baton, [MarshalAs (UnmanagedType.LPStr)] string realm, [MarshalAs (UnmanagedType.LPStr)] string user_name, [MarshalAs (UnmanagedType.Bool)] bool may_save, IntPtr pool);
 		public delegate IntPtr svn_auth_username_prompt_func_t (ref IntPtr cred, IntPtr baton, [MarshalAs (UnmanagedType.LPStr)] string realm, [MarshalAs (UnmanagedType.Bool)] bool may_save, IntPtr pool);
-		public delegate IntPtr svn_auth_ssl_server_trust_prompt_func_t (ref IntPtr cred,IntPtr baton,[MarshalAs (UnmanagedType.LPStr)] string realm,uint failures,ref svn_auth_ssl_server_cert_info_t cert_info,[MarshalAs (UnmanagedType.Bool)] bool may_save,IntPtr pool);
+		public delegate IntPtr svn_auth_ssl_server_trust_prompt_func_t (ref IntPtr cred, IntPtr baton, [MarshalAs (UnmanagedType.LPStr)] string realm, UInt32 failures, ref svn_auth_ssl_server_cert_info_t cert_info,[MarshalAs (UnmanagedType.Bool)] bool may_save, IntPtr pool);
 
-		public delegate IntPtr svn_auth_ssl_client_cert_prompt_func_t (ref IntPtr cred,IntPtr baton,[MarshalAs (UnmanagedType.LPStr)] string realm,[MarshalAs (UnmanagedType.Bool)] bool may_save,IntPtr pool);
+		public delegate IntPtr svn_auth_ssl_client_cert_prompt_func_t (ref IntPtr cred, IntPtr baton, [MarshalAs (UnmanagedType.LPStr)] string realm, [MarshalAs (UnmanagedType.Bool)] bool may_save, IntPtr pool);
 
-		public delegate IntPtr svn_auth_ssl_client_cert_pw_prompt_func_t (ref IntPtr cred,IntPtr baton,[MarshalAs (UnmanagedType.LPStr)] string realm,[MarshalAs (UnmanagedType.Bool)] bool may_save,IntPtr pool);
+		public delegate IntPtr svn_auth_ssl_client_cert_pw_prompt_func_t (ref IntPtr cred, IntPtr baton, [MarshalAs (UnmanagedType.LPStr)] string realm, [MarshalAs (UnmanagedType.Bool)] bool may_save, IntPtr pool);
 
 		public delegate IntPtr svn_log_message_receiver_t (IntPtr baton,IntPtr apr_hash_changed_paths,
-														   // TODO: See if IntPtr
-		                                                   svn_revnum_t revision,IntPtr author,IntPtr date,
-		                                                   IntPtr message,IntPtr pool);
+		                                                   svn_revnum_t revision, IntPtr author, IntPtr date,
+		                                                   IntPtr message, IntPtr pool);
 
-		public delegate IntPtr svn_readwrite_fn_t (IntPtr baton,IntPtr data,ref size_t len);
+		public delegate IntPtr svn_readwrite_fn_t (IntPtr baton, IntPtr data, ref size_t len);
 
-		public delegate void svn_wc_notify_func_t (IntPtr baton,IntPtr path,NotifyAction action,svn_node_kind_t kind,
-		                                           IntPtr mime_type,NotifyState content_state,NotifyState prop_state,
+		public delegate void svn_wc_notify_func_t (IntPtr baton, IntPtr path, NotifyAction action, svn_node_kind_t kind,
+		                                           IntPtr mime_type, NotifyState content_state, NotifyState prop_state,
 		                                           svn_revnum_t revision);
 
 		public delegate void svn_wc_notify_func2_t (IntPtr baton, ref svn_wc_notify_t notify, IntPtr pool);
